sets 'ProducesResponseType' for controller methods
