+++ Controllers/RealtorController.cs
@@ -1,4 +1,4 @@
-using real_state_web_api.Models.EntityModels;
+using real_state_web_api.Models.Entities;
 
 namespace real_state_web_api.Controllers;
