+++ Controllers/StandardController.cs
@@ -1,5 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
-using real_state_web_api.Models.EntityModels;
+using real_state_web_api.Models.Entities;
 
 namespace real_state_web_api.Controllers;
 
@@ -8,6 +8,9 @@ namespace real_state_web_api.Controllers;
 public abstract class StandardController<T> : ControllerBase
     where T : EntityModel, new()
 {
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
     [HttpPost]
     [Route("")]
     public async Task<ActionResult> Create([FromBody] T model)
@@ -19,6 +22,7 @@ public abstract class StandardController<T> : ControllerBase
         return Created($"/{result.Id}", result);
     }
 
+    [ProducesResponseType(StatusCodes.Status200OK)]
     [HttpGet]
     [Route("")]
     public async Task<ActionResult> Retrieve()
@@ -27,6 +31,8 @@ public abstract class StandardController<T> : ControllerBase
         return Ok(new List<T> { new T(), new T(), new T() });
     }
 
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
     [HttpGet]
     [Route("{id}")]
     public async Task<ActionResult> RetrieveById([FromRoute] string id)
@@ -35,6 +41,9 @@ public abstract class StandardController<T> : ControllerBase
         return Ok(new T { Id = id });
     }
 
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
     [HttpPut]
     [Route("")]
     public async Task<ActionResult> Update([FromBody] T model)
@@ -46,6 +55,8 @@ public abstract class StandardController<T> : ControllerBase
         return Ok(result);
     }
 
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
     [HttpDelete]
     [Route("{id}")]
     public async Task<ActionResult> Delete([FromRoute] string id)
