+++ Models/Entities/Rental.cs
@@ -1,0 +1,6 @@
+namespace real_state_web_api.Models.Entities;
+
+public class Rental : EntityModel
+{
+
+}
