+++ Models/Entities/Tenant.cs
@@ -1,0 +1,6 @@
+namespace real_state_web_api.Models.Entities;
+
+public class Tenant : EntityModel
+{
+
+}
