--- Models/EntityModels/Rental.cs
@@ -1,6 +1,0 @@
-namespace real_state_web_api.Models.EntityModels;
-
-public class Rental : EntityModel
-{
-
-}
